+++ src/test/java/net/oneandone/lavender/cli/LavenderIT.java
@@ -65,7 +65,7 @@ public class LavenderIT {
 
         world = World.create(withSsh);
         target = world.guessProjectHome(getClass()).join("target");
-        properties = HostProperties.load(world.guessProjectHome(getClass()).join("src/test/config/system.properties"), withSsh);
+        properties = HostProperties.load(world.guessProjectHome(getClass()).join("src/test/config/host.properties"), withSsh);
         properties.initTemp(target.join("ittemp"));
         System.out.println(name + " started: ");
         src = world.file(warFile);