+++ src/test/java/net/oneandone/lavender/filter/LavenderTest.java
@@ -68,8 +68,6 @@ public class LavenderTest {
 
         when(filterConfig.getServletContext()).thenReturn(servletContext);
         when(servletContext.getRealPath("")).thenReturn(lavenderRoot.getAbsolute());
-
-        System.setProperty("src/test/config/system.properties", WORLD.guessProjectHome(getClass()).join("src/test/config/system.properties").getAbsolute());
     }
 
     @After